Add assembly jig for OSBMX 2.4mm version
This is an assembly jig for the OSBMX 2.4mm version boards.  It lets you easily solder on the 2.4mm prongs as well as the hot swap socket.

Please note that these do not do well printed in PLA and will melt during soldering.  

It is recommended to print in something like resin but please note that a resin print has not been tested.